+++ config.py
@@ -1,0 +1,7 @@
+import os
+# Paths of projects
+CURENT_PATH = project_root = os.path.dirname(os.path.abspath(__file__))
+PROJECT_ROOT = os.path.dirname(CURENT_PATH)
+TEST_DATA_PATH = file_path = os.path.join(PROJECT_ROOT,"tests", "data")
+VERSION = "1.0.0"
+APP_NAME = "Python Vulnerability Tracker"
